refactor(wallet): extract useTokenHoldings and useWalletTransactions hooks
